Split Task and Shift into separate components, need to re-integrate data into new format
